migrate filters and icon utils
